requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.21
SQLAlchemy==2.0.36
pydantic==2.9.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
psycopg[binary]==3.1.18
python-dotenv==1.0.1
